test(rivetkit-core): move private tests to crate root
