+++ rivetkit-rust/packages/rivetkit-core/Cargo.toml
@@ -5,6 +5,7 @@ authors.workspace = true
 license.workspace = true
 edition.workspace = true
 workspace = "../../../"
+autotests = false
 
 [features]
 default = []
