+++ rivetkit-rust/packages/rivetkit-core/src/actor/metrics.rs
@@ -466,39 +466,7 @@ where
 	}
 }
 
+// Test shim keeps moved tests in crate-root tests/ with private-module access.
 #[cfg(test)]
-mod tests {
-	use std::panic::{AssertUnwindSafe, catch_unwind};
-
-	use super::*;
-
-	#[test]
-	fn duplicate_metric_registration_uses_noop_fallback() {
-		let registry = Registry::new();
-		let first = IntGauge::with_opts(Opts::new(
-			"duplicate_actor_metric",
-			"first duplicate metric",
-		))
-		.expect("first gauge should be valid");
-		let second = IntGauge::with_opts(Opts::new(
-			"duplicate_actor_metric",
-			"second duplicate metric",
-		))
-		.expect("second gauge should be valid");
-
-		register_metric(&registry, first.clone());
-		let result = catch_unwind(AssertUnwindSafe(|| {
-			register_metric(&registry, second.clone());
-		}));
-
-		assert!(result.is_ok());
-		assert_eq!(
-			1,
-			registry
-				.gather()
-				.iter()
-				.filter(|family| family.name() == "duplicate_actor_metric")
-				.count()
-		);
-	}
-}
+#[path = "../../tests/metrics.rs"]
+mod tests;
