+++ rivetkit-rust/packages/rivetkit-core/src/actor/work_registry.rs
@@ -120,39 +120,7 @@ impl Drop for RegionGuard {
 /// `CountGuard` is the same RAII shape as `RegionGuard`, but used for task-counting sites.
 pub(crate) type CountGuard = RegionGuard;
 
+// Test shim keeps moved tests in crate-root tests/ with private-module access.
 #[cfg(test)]
-mod tests {
-	use std::panic::{AssertUnwindSafe, catch_unwind};
-
-	use super::WorkRegistry;
-
-	#[test]
-	fn region_guard_drop_decrements_counter() {
-		let work = WorkRegistry::new();
-		assert_eq!(work.keep_awake.load(), 0);
-
-		{
-			let _guard = work.keep_awake_guard();
-			assert_eq!(work.keep_awake.load(), 1);
-		}
-
-		assert_eq!(work.keep_awake.load(), 0);
-	}
-
-	#[test]
-	fn region_guard_drop_during_panic_unwind_decrements_counter() {
-		let work = WorkRegistry::new();
-
-		let result = catch_unwind(AssertUnwindSafe(|| {
-			let _guard = work.keep_awake_guard();
-			assert_eq!(work.keep_awake.load(), 1);
-			panic!("boom");
-		}));
-
-		assert!(
-			result.is_err(),
-			"panic should propagate through catch_unwind"
-		);
-		assert_eq!(work.keep_awake.load(), 0);
-	}
-}
+#[path = "../../tests/work_registry.rs"]
+mod tests;
