+++ rivetkit-rust/packages/rivetkit-core/src/registry/envoy_callbacks.rs
@@ -294,60 +294,7 @@ fn preloaded_kv_from_protocol(preloaded_kv: protocol::PreloadedKv) -> PreloadedK
 	)
 }
 
+// Test shim keeps moved tests in crate-root tests/ with private-module access.
 #[cfg(test)]
-mod preload_tests {
-	use super::*;
-	use crate::actor::state::{PersistedActor, encode_persisted_actor};
-
-	#[test]
-	fn decode_preloaded_persisted_actor_distinguishes_bundle_states() {
-		assert_eq!(
-			decode_preloaded_persisted_actor(None).expect("no bundle should decode"),
-			PreloadedPersistedActor::NoBundle
-		);
-
-		let requested_empty = protocol::PreloadedKv {
-			entries: Vec::new(),
-			requested_get_keys: vec![PERSIST_DATA_KEY.to_vec()],
-			requested_prefixes: Vec::new(),
-		};
-		assert_eq!(
-			decode_preloaded_persisted_actor(Some(&requested_empty))
-				.expect("empty bundle should decode"),
-			PreloadedPersistedActor::BundleExistsButEmpty
-		);
-
-		let not_requested = protocol::PreloadedKv {
-			entries: Vec::new(),
-			requested_get_keys: Vec::new(),
-			requested_prefixes: Vec::new(),
-		};
-		assert_eq!(
-			decode_preloaded_persisted_actor(Some(&not_requested))
-				.expect("unrequested bundle should decode"),
-			PreloadedPersistedActor::NoBundle
-		);
-
-		let persisted = PersistedActor {
-			state: vec![1, 2, 3],
-			..PersistedActor::default()
-		};
-		let with_actor = protocol::PreloadedKv {
-			entries: vec![protocol::PreloadedKvEntry {
-				key: PERSIST_DATA_KEY.to_vec(),
-				value: encode_persisted_actor(&persisted).expect("persisted actor should encode"),
-				metadata: protocol::KvMetadata {
-					version: Vec::new(),
-					update_ts: 0,
-				},
-			}],
-			requested_get_keys: vec![PERSIST_DATA_KEY.to_vec()],
-			requested_prefixes: Vec::new(),
-		};
-		assert_eq!(
-			decode_preloaded_persisted_actor(Some(&with_actor))
-				.expect("persisted actor bundle should decode"),
-			PreloadedPersistedActor::Some(persisted)
-		);
-	}
-}
+#[path = "../../tests/envoy_callbacks.rs"]
+mod tests;
