+++ rivetkit-rust/packages/rivetkit-core/src/websocket.rs
@@ -227,6 +227,7 @@ impl fmt::Debug for WebSocket {
 	}
 }
 
+// Test shim keeps moved tests in crate-root tests/ with private-module access.
 #[cfg(test)]
-#[path = "../tests/modules/websocket.rs"]
+#[path = "../tests/websocket.rs"]
 mod tests;
